Reduce duplication by extracting common update function of InstantiatedTypes for PartialResults.
